Definitions of microSLAM Library
The name of the new SLAM Library shall be "microSLAM".
CoreSLAM is still contained, but not included. The first files of microslam were included and replaced the CoreSLAM definitions. The datastructs and containers of the SLAM Lib were created. The map is ready and can be displayed by a function as an area (as fast as possible).