Implementation of distanceScanToMap, comm interface
Implementated tinySLAM ts_distance_scan_to_map. It calculates the ambiguity of the newest scan with the scan on a given position and returns a number proportional to the ambiguity.
Also added files for the comm interface (communication to sub controller) and the initialization of USART3 (PB10/11). Raw communication with sub controller works.