Implementation of subcontroller communication protocol
Communication to subcontroller now completely works (object orientated). You can define if you want to read or write from/to the controller, define how many bytes you want to write/read and in which register you want to begin.
Next step is a reliable "watchdog" of the communication and functions using the protocol (like "readSensors" or "setSpeed").