Major commit: Functionality to stream the map to the PC
Now, the robot is able to stream the map to the PC basis station. Therefore, we increased the baud rate to 460800.
The map and the robot position are send to the PC in the debug task. The protocol looks like this:

Protocol for PC User interface
[Startseq][Length][Checksum][ID][Data]

[Startseq]: ["PCUI_MSG"] (8 chars)
[Lenght]: [{b2},{b1_lsb}] (16bit; 2 chars)
[checksum]: (Sum of all Data chars) [{b4},{b3},{b2},{b1_lsb}] (32bit; 4 chars)
[ID]: (3 chars)
		["MPD"]: Map Data (13 chars). Rob->PC
					- resolution (mm) (1byte)
					- size x (2byte)
					- size y "
					- size z (1byte)
					- rob x  (2byte)
					- rob y  "
					- rob z  (1byte)
					- dir    (2byte)
		["MAP"]: Map transmission (size y + 3 chars). Rob->PC
					- current stage (1byte)
					- current line (transmission linewise) (2byte)
					- Pixel (size x byte)
		["WAY"]: Waypoints (n waypoints * 7 chars). Bidirectional
					- ID (1byte)
					- x  (2byte)
					- y  "
					- z  "
		["STA"]: Status (1 char). Bidirectional.
					- start/stop
	[Data]: [Lenght] chars

A function "puts_l" was added to printf. It can send a string with a defined length (usually, a streng ends with \0, but we see 0 sometimes as data, so actually, we don’t send a string, but a character array (the message)).
A function "pcui_sendMap" manages all the protocol stuff and cuts the map into piecec to stream it to the pc. It only needs the slam container.
The function bases on "pcui_sendMsg", which needs the message id, the length and the message.